Add Codacy badge to README.md (#15)
* Add Codacy badge to README.md

Introduced a Codacy badge to the README.md file to display the project's code quality grade. The badge links to the project's dashboard on Codacy for more detailed information.

* Remove Codacy Badge from README.md

The Codacy Badge, which provided a visual representation of the project's code quality grade, has been removed from the README.md file. This badge linked to the Codacy dashboard for the project.

---------

Co-authored-by: AlaBakhamis <32382009+AlaBakhamis@users.noreply.github.com>